Allow only disabled negotiator test; Move utils test functions to separate module
