+++ agent/provider/tests/utils/mod.rs
@@ -1,0 +1,2 @@
+#[allow(dead_code)]
+pub mod rules;
